pyproject.toml
[tool.poetry]
name = "cryptopals"
version = "0.1.0"
description = "CryptoPals solutions"
authors = ["Zhenkai Weng <theoaktree1040@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.17"
Flask = "^2.2.3"
requests = "^2.28.2"
pwntools = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
